backend/requirements.txt
cffi==1.13.2
click==7.1.2
dataclasses==0.6
ecdsa==0.14.1
fastapi==0.60.1
flake8==3.8.3
h11==0.9.0
importlib-metadata==1.7.0
mccabe==0.6.1
passlib==1.7.2
pyasn1==0.4.8
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
pymongo==3.11.0
python-jose==3.2.0
readline==6.2.4.1
rsa==4.6
six==1.15.0
starlette==0.13.6
uvicorn==0.11.8
websockets==8.1
zipp==3.1.0
